Reduce indirections
